requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
asttokens==2.4.1
attrs==25.4.0
Authlib==1.6.5
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==46.0.2
cyclopts==3.24.0
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.8
Django==5.1.7
django-cors-headers==4.7.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.1.0
fastmcp==2.12.4
feedparser==6.0.12
filelock==3.16.1
flatbuffers==25.2.10
git-filter-repo==2.45.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
isodate==0.7.2
jedi==0.19.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mcp==1.17.0
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.1
oauthlib==3.2.2
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.27.1
protobuf==5.29.6
psutil==6.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.4
pyperclip==1.11.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-stdnum==1.19
python-ulid==1.1.0
pytz==2023.3.post1
pywin32==311
PyYAML==6.0.2
pyzmq==26.2.0
redis==4.6.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.1
rjsmin==1.2.0
rpds-py==0.27.1
rsa==4.9
setuptools==75.6.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.11.0
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.2.0
urllib3==1.26.16
uvicorn==0.37.0
hypercorn>=0.17.0
virtualenv==20.26.6
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.1.1
xlwt==1.3.0
zipp==3.17.0
